Introduce Session to drive job execution.
