+++ tests/benchmark/test_bad_setup.py
@@ -1,4 +1,3 @@
-import os
 from unittest.mock import patch
 
 import pytest
@@ -7,9 +6,11 @@ from gretel_trainer.benchmark import GretelLSTM, compare
 from gretel_trainer.benchmark.core import BenchmarkConfig, BenchmarkException
 
 
-def test_bad_session_exits_early(iris):
+def test_bad_session_exits_early(iris, tmp_path):
     class SomeException(Exception):
         pass
+
+    should_not_be_created = tmp_path / "should_not_be_created"
 
     with patch(
         "gretel_trainer.benchmark.comparison.configure_session"
@@ -21,32 +22,36 @@ def test_bad_session_exits_early(iris):
                 datasets=[iris],
                 models=[GretelLSTM],
                 config=BenchmarkConfig(
-                    working_dir="should_not_be_created",
+                    working_dir=should_not_be_created,
                 ),
             )
 
-    assert not os.path.exists("should_not_be_created")
+    assert not should_not_be_created.exists()
 
 
-def test_dataset_names_must_be_unique(iris):
+def test_dataset_names_must_be_unique(iris, tmp_path):
+    should_not_be_created = tmp_path / "should_not_be_created"
+
     with pytest.raises(BenchmarkException):
         compare(
             datasets=[iris, iris],
             models=[GretelLSTM],
             config=BenchmarkConfig(
-                working_dir="should_not_be_created",
+                working_dir=should_not_be_created,
             ),
         )
-    assert not os.path.exists("should_not_be_created")
+    assert not should_not_be_created.exists()
 
 
-def test_model_names_must_be_unique(iris):
+def test_model_names_must_be_unique(iris, tmp_path):
+    should_not_be_created = tmp_path / "should_not_be_created"
+
     with pytest.raises(BenchmarkException):
         compare(
             datasets=[iris],
             models=[GretelLSTM, GretelLSTM],
             config=BenchmarkConfig(
-                working_dir="should_not_be_created",
+                working_dir=should_not_be_created,
             ),
         )
-    assert not os.path.exists("should_not_be_created")
+    assert not should_not_be_created.exists()
